# docs/openapi.yaml
openapi: 3.0.3
info:
  title: FitSync AI API
  version: 2.0.0
servers:
  - url: /api
paths:
...
  /food/search:
    get:
      summary: Food search
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Food results
  /food/barcode/{barcode}:
    get:
      summary: Barcode lookup
      parameters:
        - name: barcode
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Food item
  /food/details/{fdcId}:
    get:
      summary: USDA Food details
      parameters:
        - name: fdcId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Food item
...